pyproject.toml
[build-system]
requires = ["setuptools>=70.0", "wheel>=0.42", "pip>=24.0"]

[project]
name = "cv-generator"
version = "0.1.0"
description = "An AI-powered CV generator using Streamlit and OpenAI"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "htbuilder>=0.9.0",
    "streamlit>=1.49.0",
    "openai>=0.27.0",
    "PyMuPDF>=1.26.0",
    "python-docx>=1.1.0",
]
